Accept new values from mutate_json_record
The current implementation modifies an argument which is not always good. The
assumption that the extra argument is not re-used by the callee is reasonable
but ultimately cannot be relied on all the time. This change gives implementers
the change to catch an extra dict that contains references to re-used objects
and replace them with their own objects. The current implementation will not
take care of that for them but enables the creation of such a function.

We also retain backwards compatibility by using the modified argument if no
return value was provided.